requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=h0e60522_0
aiofiles=22.1.0=py38haa95532_0
aiosqlite=0.18.0=py38haa95532_0
altair=5.0.1=pyhd8ed1ab_1
anyio=3.5.0=py38haa95532_0
appdirs=1.4.4=pyhd3eb1b0_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
arrow-cpp=11.0.0=ha81ea56_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py38haa95532_0
aws-c-common=0.6.8=h2bbff1b_1
aws-c-event-stream=0.1.6=hd77b12b_6
aws-checksums=0.1.11=h2bbff1b_2
aws-sdk-cpp=1.8.185=hd77b12b_1
babel=2.11.0=py38haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py38haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.2=pyhd8ed1ab_0
boost-cpp=1.78.0=h5b4e17d_0
bottleneck=1.3.5=py38h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py38h2bbff1b_1003
bzip2=1.0.8=h8ffe710_4
c-ares=1.19.0=h2bbff1b_0
ca-certificates=2023.7.22=h56e8100_0
cachetools=5.3.1=pyhd8ed1ab_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38haa95532_0
colorama=0.4.6=py38haa95532_0
comm=0.1.2=py38haa95532_0
contourpy=1.0.5=py38h59b6b97_0
cryptography=41.0.2=py38hac1b9e3_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38haa95532_0
executing=0.8.3=pyhd3eb1b0_0
faker=18.9.0=py38haa95532_0
filelock=3.9.0=py38haa95532_0
flask=2.2.2=py38haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
fsspec=2023.4.0=py38haa95532_0
gflags=2.2.2=ha925a31_1004
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.10=pyhd8ed1ab_0
gitpython=3.1.32=pyhd8ed1ab_0
glib=2.69.1=h5dc1a3c_2
glog=0.5.0=h4797de2_0
greenlet=2.0.1=py38hd77b12b_0
grpc-cpp=1.48.2=hfe90ff0_1
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py38haa95532_0
importlib-metadata=6.0.0=py38haa95532_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.25.0=py38h9909e9c_0
ipython=8.12.2=py38haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.4=py38haa95532_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jedi=0.18.1=py38haa95532_1
jinja2=3.1.2=py38haa95532_0
joblib=1.2.0=py38haa95532_0
jpeg=9e=h2bbff1b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.17.3=py38haa95532_0
jupyter=1.0.0=py38haa95532_8
jupyter_client=7.4.9=py38haa95532_0
jupyter_console=6.6.3=py38haa95532_0
jupyter_core=5.3.0=py38haa95532_0
jupyter_events=0.6.3=py38haa95532_0
jupyter_server=1.23.4=py38haa95532_0
jupyter_server_fileid=0.9.0=py38haa95532_0
jupyter_server_ydoc=0.8.0=py38haa95532_1
jupyter_ydoc=0.2.4=py38haa95532_0
jupyterlab=3.6.3=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.22.0=py38haa95532_0
jupyterlab_widgets=3.0.5=py38haa95532_0
kiwisolver=1.4.4=py38hd77b12b_0
krb5=1.20.1=h5b6d351_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcurl=8.1.1=h86230a5_0
libdeflate=1.17=h2bbff1b_0
libevent=2.1.10=h91dae50_4
libffi=3.4.4=hd77b12b_0
libiconv=1.16=h2bbff1b_2
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=he2ea4bf_2
libthrift=0.15.0=h4364b78_2
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.10.4=h0ad7f3c_1
libxslt=1.1.37=h2bbff1b_1
lightning-utilities=0.7.1=py38haa95532_1
lxml=4.9.2=py38h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.7.1=py38haa95532_1
matplotlib-base=3.7.1=py38hf11a4ad_1
matplotlib-inline=0.1.6=py38haa95532_0
mdurl=0.1.0=pyhd8ed1ab_0
mistune=0.8.4=py38he774522_1000
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.6=py38hf11a4ad_1
mkl_random=1.2.2=py38hf11a4ad_1
mpmath=1.3.0=py38haa95532_0
munkres=1.1.4=py_0
nbclassic=0.5.5=py38haa95532_0
nbclient=0.5.13=py38haa95532_0
nbconvert=6.5.4=py38haa95532_0
nbformat=5.7.0=py38haa95532_0
nest-asyncio=1.5.6=py38haa95532_0
networkx=3.1=py38haa95532_0
ninja=1.10.2=haa95532_5
ninja-base=1.10.2=h6d14046_5
notebook=6.5.4=py38haa95532_1
notebook-shim=0.2.2=py38haa95532_0
numexpr=2.8.4=py38h7b80656_1
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
openssl=3.0.10=h2bbff1b_2
orc=1.7.4=h623e30f_1
packaging=23.0=py38haa95532_0
pandas=2.0.3=py38h4ed8f06_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py38hd77b12b_0
pip=23.2.1=py38haa95532_0
pkgutil-resolve-name=1.3.10=py38haa95532_0
platformdirs=3.10.0=py38haa95532_0
plotly=5.9.0=py38haa95532_0
ply=3.11=py38_0
pooch=1.4.0=pyhd3eb1b0_0
prometheus_client=0.14.1=py38haa95532_0
prompt-toolkit=3.0.36=py38haa95532_0
prompt_toolkit=3.0.36=hd3eb1b0_0
protobuf=3.20.3=py38haa244fe_1
psutil=5.9.0=py38h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=11.0.0=py38haa45b5f_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.0=pyhd8ed1ab_0
pygments=2.15.1=py38haa95532_1
pympler=1.0.1=pyhd8ed1ab_0
pyopenssl=23.2.0=py38haa95532_0
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
pyrsistent=0.18.0=py38h196d8e1_0
pysocks=1.7.1=py38haa95532_0
python=3.8.17=h1aa4202_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py38haa95532_0
python-json-logger=2.0.7=py38haa95532_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.0.1=cpu_py38hb0bdfb8_0
pytorch-lightning=2.0.3=pyhd8ed1ab_0
pytz=2022.7=py38haa95532_0
pywin32=305=py38h2bbff1b_0
pywinpty=2.0.10=py38h5da7b33_0
pyyaml=6.0=py38h2bbff1b_1
pyzmq=23.2.0=py38hd77b12b_0
qt-main=5.15.2=h879a1e9_9
qt-webengine=5.15.9=h5bd16bc_7
qtconsole=5.4.2=py38haa95532_0
qtpy=2.2.0=py38haa95532_0
qtwebkit=5.212=h2bbfb41_5
re2=2022.04.01=h0e60522_0
requests=2.31.0=py38haa95532_0
rfc3339-validator=0.1.4=py38haa95532_0
rfc3986-validator=0.1.1=py38haa95532_0
rich=13.5.1=pyhd8ed1ab_0
scikit-learn=1.2.2=py38hd77b12b_1
scipy=1.10.1=py38hdcfc7df_1
seaborn=0.12.2=py38haa95532_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=68.0.0=py38haa95532_0
sip=6.6.2=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smmap=3.0.5=pyh44b312d_0
snappy=1.1.9=h6c2663c_0
sniffio=1.2.0=py38haa95532_1
soupsieve=2.4=py38haa95532_0
sqlalchemy=1.4.39=py38h2bbff1b_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.26.0=pyhd8ed1ab_0
sympy=1.11.1=py38haa95532_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.2=py38haa95532_0
terminado=0.17.1=py38haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py38haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py38haa95532_0
toolz=0.12.0=pyhd8ed1ab_0
torchmetrics=0.11.4=py38h9909e9c_1
torchvision=0.13.1=cpu_py38h378ed51_0
tornado=6.3.2=py38h2bbff1b_0
tqdm=4.65.0=py38hd4e2768_0
traitlets=5.7.1=py38haa95532_0
typing-extensions=4.7.1=py38haa95532_0
typing_extensions=4.7.1=py38haa95532_0
tzlocal=2.1=pyh9f0ad1d_0
urllib3=1.26.16=py38haa95532_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.21.2=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=py38haa244fe_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38haa95532_4
werkzeug=2.2.3=py38haa95532_0
wheel=0.38.4=py38haa95532_0
widgetsnbextension=4.0.5=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
winpty=0.4.3=4
wtforms=2.3.3=pyhd3eb1b0_0
xz=5.4.2=h8cc25b3_0
y-py=0.5.9=py38hb6bf4ef_0
yaml=0.2.5=he774522_0
ypy-websocket=0.8.2=py38haa95532_0
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
